Finished models
CSV files are handled through spreadsheets.py, which serves as a sort
of intermediary to a database, whose structure can be substantially
modified and adapted without breaking any other part of the program.

countryYear.py serves as a more abstract module through which the rest
of the program can store and access Location objects without worrying
about the implementation details of how those Location objects are
accessed or stored.

- pk’s start at 2 and count up from there. Locations stored at pk’s may
be modified but pk’s may not be vacated.